+++ tests/test_add_agents_operations_bulk.py
@@ -294,6 +294,8 @@ class TestAddOperationsBulkSomeFailure(unittest.TestCase):
         self.assertEqual(resp[0].get("id"), self.agent_id)
         self.assertTrue("message" in resp[0].keys())
         self.assertFalse("error" in resp[0].keys())
-        self.assertTrue(len(resp) == 1)
+        self.assertTrue(len(resp) == len(payload))
+        for i in range(len(invalid_data.UNDEFINED_KEY)):
+            self.assertTrue("error" in resp[i + 1].keys())
 
         self.addCleanup(self.clean_up_agent, self.agent_id)